feat(wintertc): migrate timer APIs from boa_runtime
Move setTimeout/clearTimeout/setInterval/clearInterval from
boa_runtime::interval into boa_wintertc::timers, following the
move-and-re-export pattern used for atob/btoa (#5418). The module is renamed
interval -> timers to match its TC55 category.

- core/runtime/src/interval.rs and its tests are deleted; the code now lives
  only in boa_wintertc::timers.
- boa_runtime re-exports it under the historical name
  (pub use boa_wintertc::timers as interval) so the public path
  boa_runtime::interval (used by e.g. the WPT harness's clear_all) and the
  register(context) signature are unchanged.
- TimeoutExtension delegates to boa_wintertc::timers::register.
- boa_wintertc mirrors boa_runtime's test-only
  allow(clippy::needless_raw_string_hashes); runtime's now-unused
  inspect_context_async test helper gains allow(unused).
